Add TABPY_EVALUATE_WITH configuration parameter
