+++ .vscode/settings.json
@@ -18,5 +18,6 @@
     "python.testing.unittestEnabled": false,
     "python.testing.nosetestsEnabled": false,
     "python.testing.pytestEnabled": true,
-    "python.linting.pycodestyleEnabled": true
+    "python.linting.pycodestyleEnabled": true,
+    "python.pythonPath": "c:\\Users\\ogolovatyi\\.conda\\envs\\Python 3.7 (tornado 6)\\python.exe"
 }
