+++ CHANGELOG
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.2.0
+
+### Improvements
+
+- Added support for Rserve protocol.
+
 ## v1.1.0
 
 ### Improvements
