+++ tabpy/tabpy_server/handlers/base_handler.py
@@ -4,7 +4,7 @@ import concurrent
 import json
 import logging
 import tornado.web
-from tabpy.tabpy_server.app.SettingsParameters import SettingsParameters
+from tabpy.tabpy_server.app.app_parameters import SettingsParameters
 from tabpy.tabpy_server.handlers.util import hash_password
 import uuid
