+++ tabpy/tabpy_server/handlers/management_handler.py
@@ -6,7 +6,7 @@ from uuid import uuid4 as random_uuid
 
 from tornado import gen
 
-from tabpy.tabpy_server.app.SettingsParameters import SettingsParameters
+from tabpy.tabpy_server.app.app_parameters import SettingsParameters
 from tabpy.tabpy_server.handlers import MainHandler
 from tabpy.tabpy_server.handlers.base_handler import STAGING_THREAD
 from tabpy.tabpy_server.management.state import get_query_object_path
