+++ tabpy/tabpy_server/handlers/upload_destination_handler.py
@@ -1,4 +1,4 @@
-from tabpy.tabpy_server.app.SettingsParameters import SettingsParameters
+from tabpy.tabpy_server.app.app_parameters import SettingsParameters
 from tabpy.tabpy_server.handlers import ManagementHandler
 import os
